refactor: change class path
