+++ src/main/java/exception/ConnectionClosedException.java
@@ -1,4 +1,4 @@
-package controller;
+package exception;
 
 public class ConnectionClosedException extends RuntimeException {
     public ConnectionClosedException(Throwable cause) {
